This fix makes sure that python code can query correctly if oneDNN is enabled or not
