+++ tensorflow/core/util/onednn_env_vars.h
@@ -20,6 +20,7 @@ limitations under the License.
 namespace tensorflow {
 bool AreWeightsFrozen();
 bool UseSystemAlloc();
+bool IsMKLEnabled();
 }  // namespace tensorflow
 #endif  // INTEL_MKL
 #endif  // TENSORFLOW_CORE_UTIL_ONEDNN_ENV_VARS_H_
