+++ tensorflow/core/util/port.cc
@@ -14,7 +14,7 @@ limitations under the License.
 ==============================================================================*/
 
 #include "tensorflow/core/util/port.h"
-
+#include "tensorflow/core/util/onednn_env_vars.h"
 
 namespace tensorflow {
 
@@ -59,7 +59,5 @@ bool GpuSupportsHalfMatMulAndConv() {
 #endif
 }
 
-bool IsMklEnabled() {
-  return IsMKLEnabled();
-}
+bool IsMklEnabled() { return IsMKLEnabled(); }
 }  // end namespace tensorflow
